Fix up plugin and tool priority code:
  * Priority is now a double instead of an int, similar to the
    MenuEntry's weight parameter. This makes it essentially always
    possible to squeeze a new processor or tool between two others in
    terms of priority.
  * Two new priority presets ("very high" and "very low") now exist.
  * The priorities of several processor plugins and tools have been
    adjusted.
  * The "first" and "last" priorities are now really always first and
    last, through the use of negative and positive infinity
    respectively. It is no longer possible to insert a processor before
    one with first priority, or after one with last priority.
  * The javadoc for several priority-related elements has been updated.

This used to be revision r4401.

+++ core/ext/src/main/java/imagej/ext/UIDetails.java
@@ -61,7 +61,7 @@ public interface UIDetails extends BasicDetails, Comparable<UIDetails> {
 	String getIconPath();
 
 	/** Gets the sort priority of the object. */
-	int getPriority();
+	double getPriority();
 
 	/**
 	 * Gets whether the object can be selected (e.g., checking and unchecking its
@@ -91,7 +91,7 @@ public interface UIDetails extends BasicDetails, Comparable<UIDetails> {
 	void setIconPath(String iconPath);
 
 	/** Sets the sort priority of the object. */
-	void setPriority(int priority);
+	void setPriority(double priority);
 
 	/** Sets whether the object should be enabled in the user interface. */
 	void setEnabled(boolean enabled);
